requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
streamlit>=1.30
python-dotenv>=1.0

